+++ components/bridge/dao/impl/src/main/java/com/flipperdevices/bridge/dao/impl/api/key/SimpleKeyApiImpl.kt
@@ -1,6 +1,5 @@
 package com.flipperdevices.bridge.dao.impl.api.key
 
-import com.flipperdevices.bridge.dao.api.delegates.FavoriteApi
 import com.flipperdevices.bridge.dao.api.delegates.key.SimpleKeyApi
 import com.flipperdevices.bridge.dao.api.model.FlipperFileType
 import com.flipperdevices.bridge.dao.api.model.FlipperKey
@@ -21,13 +20,11 @@ import kotlinx.coroutines.withContext
 
 @ContributesBinding(AppGraph::class, SimpleKeyApi::class)
 class SimpleKeyApiImpl @Inject constructor(
-    keysDaoProvider: Provider<SimpleKeyDao>,
-    favoriteApiProvider: Provider<FavoriteApi>
+    keysDaoProvider: Provider<SimpleKeyDao>
 ) : SimpleKeyApi, LogTagProvider {
     override val TAG = "SimpleKeyApi"
 
     private val simpleKeyDao by keysDaoProvider
-    private val favoriteApi by favoriteApiProvider
 
     override suspend fun getAllKeys(): List<FlipperKey> = withContext(Dispatchers.IO) {
         return@withContext simpleKeyDao.getAll().map { it.toFlipperKey() }