+++ components/core/ui/navigation/src/main/java/com/flipperdevices/core/ui/navigation/AggregateFeatureEntry.kt
@@ -1,0 +1,10 @@
+package com.flipperdevices.core.ui.navigation
+
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.NavHostController
+
+interface AggregateFeatureEntry : FeatureEntry {
+    fun NavGraphBuilder.navigation(
+        navController: NavHostController
+    )
+}