+++ components/filemanager/api/src/main/java/com/flipperdevices/filemanager/api/share/FileManagerDeepLinkHandler.kt
@@ -1,5 +1,0 @@
-package com.flipperdevices.filemanager.api.share
-
-import com.flipperdevices.deeplink.api.DeepLinkHandler
-
-interface FileManagerDeepLinkHandler : DeepLinkHandler