+++ components/filemanager/impl/src/main/java/com/flipperdevices/filemanager/impl/composable/ComposableFileManagerUploadedScreen.kt
@@ -1,0 +1,29 @@
+package com.flipperdevices.filemanager.impl.composable
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.stringResource
+import com.flipperdevices.filemanager.impl.R
+import com.flipperdevices.filemanager.impl.composable.list.ComposableFileManagerContent
+import com.flipperdevices.filemanager.impl.model.FileManagerState
+import com.flipperdevices.filemanager.sharecommon.composable.ComposableAlertDialog
+import com.flipperdevices.filemanager.sharecommon.model.ShareState
+
+@Composable
+fun ComposableFileManagerUploadedScreen(
+    fileManagerState: FileManagerState,
+    shareState: ShareState,
+    onCancel: () -> Unit
+) {
+    Box {
+        ComposableFileManagerContent(fileManagerState = fileManagerState, onFileClick = {})
+        ComposableAlertDialog(
+            title = stringResource(
+                R.string.receive_dialog_title,
+                shareState.name
+            ),
+            downloadProgress = shareState.downloadProgress,
+            onCancel = onCancel
+        )
+    }
+}