+++ components/filemanager/impl/src/main/java/com/flipperdevices/filemanager/impl/viewmodels/FileManagerViewModel.kt
@@ -5,37 +5,35 @@ import com.flipperdevices.bridge.api.model.wrapToRequest
 import com.flipperdevices.bridge.service.api.FlipperServiceApi
 import com.flipperdevices.bridge.service.api.provider.FlipperBleServiceConsumer
 import com.flipperdevices.bridge.service.api.provider.FlipperServiceProvider
-import com.flipperdevices.core.di.ComponentHolder
 import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.log.info
 import com.flipperdevices.core.ui.lifecycle.LifecycleViewModel
-import com.flipperdevices.filemanager.impl.di.FileManagerComponent
+import com.flipperdevices.filemanager.impl.api.PATH_KEY
 import com.flipperdevices.filemanager.impl.model.FileItem
 import com.flipperdevices.filemanager.impl.model.FileManagerState
 import com.flipperdevices.protobuf.main
 import com.flipperdevices.protobuf.storage.Storage
 import com.flipperdevices.protobuf.storage.listRequest
 import java.io.File
-import javax.inject.Inject
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
+import tangle.inject.TangleParam
+import tangle.viewmodel.VMInject
 
-class FileManagerViewModel(
+class FileManagerViewModel @VMInject constructor(
+    serviceProvider: FlipperServiceProvider,
+    @TangleParam(PATH_KEY)
     private val directory: String
 ) : LifecycleViewModel(), FlipperBleServiceConsumer, LogTagProvider {
     override val TAG = "FileManagerViewModel"
-
-    @Inject
-    lateinit var serviceProvider: FlipperServiceProvider
 
     private var fileManagerStateFlow =
         MutableStateFlow(FileManagerState(currentPath = directory))
 
     init {
-        ComponentHolder.component<FileManagerComponent>().inject(this)
         serviceProvider.provideServiceApi(consumer = this, lifecycleOwner = this)
     }