+++ components/filemanager/sharecommon/src/main/java/com/flipperdevices/filemanager/sharecommon/composable/ComposableAlertDialog.kt
@@ -18,6 +18,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
+import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.filemanager.sharecommon.R
 import com.flipperdevices.filemanager.sharecommon.model.DownloadProgress
 
@@ -88,7 +89,8 @@ fun ComposableFixedProgress(fixedProgress: DownloadProgress.Fixed) {
         )
         LinearProgressIndicator(
             modifier = Modifier.fillMaxWidth(),
-            progress = animatedProgress
+            progress = animatedProgress,
+            color = LocalPallet.current.accent
         )
     }
 }