+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [Feature] Add outdated application dialog
 - [Feature] Add instantly search for flipper if device already connected or bounded
 - [Feature] Try auto connect on each app open
+- [Feature] Add ability to upload file from file manager
 - [BUGFIX] Exit from emulate screen
 - [BUGFIX] Lock portrait orientation
 - [BUGFIX] Update card not shown when flipper not connected
@@ -31,6 +32,7 @@
 - [REFACTOR] Preview (Updater card/screen, Device info)
 - [Feature] Button Emulate/Send with auto close
 - [REFACTOR] Add cache for gradle build
+- [REFACTOR] Migrate FileManager to compose navigation
 
 # 1.1.5 - HotFix